+++ packages/Panel/stories/Panel.backyard.stories.js
@@ -3,6 +3,7 @@ import { getStoryName } from "storybook/storyTree";
 import { testStoryParameters } from "storybook/assets/storyParameters";
 import ExampleStory from "storybook/components/ExampleStory";
 import WithSortable from "./sandbox/WithSortable";
+import WebComponent from "./sandbox/WebComponent";
 import Panel from "../src";
 
 const storyName = getStoryName("Panel");
@@ -17,8 +18,17 @@ export const WithSortableStory = () => (
     <WithSortable />
   </ExampleStory>
 );
-
 WithSortableStory.story = {
   name: "With Sortable",
+  parameters: testStoryParameters,
+};
+
+export const WebComponentStory = () => (
+  <ExampleStory storyName="Web Component" component="Panel" fileName="sandbox/WebComponent.js">
+    <WebComponent />
+  </ExampleStory>
+);
+WebComponentStory.story = {
+  name: "Web Component",
   parameters: testStoryParameters,
 };