+++ packages/Counter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0bf34d9]
+- Updated dependencies [0dcf307]
+- Updated dependencies [9ac6aca]
+  - @paprika/helpers@2.3.0
+  - @paprika/tokens@2.0.0
+  - @paprika/stylers@1.1.6
+
 ## 1.1.5-next.2
 
 ### Patch Changes