+++ packages/DatePicker/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0bf34d9]
+- Updated dependencies [0dcf307]
+- Updated dependencies [9ac6aca]
+  - @paprika/popover@2.0.4
+  - @paprika/helpers@2.3.0
+  - @paprika/tokens@2.0.0
+  - @paprika/calendar@2.0.6
+  - @paprika/date-input@1.0.28
+  - @paprika/stylers@1.1.6
+
 ## 2.0.6-next.2
 
 ### Patch Changes