+++ packages/FormElement/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 4.0.30
+
+### Patch Changes
+
+- 7f6da83: Enable tooltip aligment as a prop and set default to top
+
+## 4.0.30-next.0
+
+### Patch Changes
+
+- 7f6da83: Enable tooltip aligment as a prop and set default to top
+
+## 4.0.29
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0bf34d9]
+- Updated dependencies [0dcf307]
+- Updated dependencies [9ac6aca]
+  - @paprika/popover@2.0.4
+  - @paprika/helpers@2.3.0
+  - @paprika/tokens@2.0.0
+  - @paprika/icon@2.1.7
+  - @paprika/stylers@1.1.6
+
 ## 4.0.29-next.2
 
 ### Patch Changes