+++ packages/ListBoxBrowser/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 1.0.42
+
+### Patch Changes
+
+- Updated dependencies [0dcf307]
+- Updated dependencies [0bf34d9]
+- Updated dependencies [0dcf307]
+- Updated dependencies [9ac6aca]
+  - @paprika/list-box@4.0.6
+  - @paprika/helpers@2.3.0
+  - @paprika/tokens@2.0.0
+  - @paprika/icon@2.1.7
+  - @paprika/raw-button@1.0.12
+  - @paprika/spinner@1.0.16
+  - @paprika/stylers@1.1.6
+
 ## 1.0.42-next.2
 
 ### Patch Changes