+++ packages/Popover/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 2.0.4
+
+### Patch Changes
+
+- 0dcf307: Refactored to use `getActiveElement` from `@paprika/helpers` to support web components.
+- Updated dependencies [0dcf307]
+- Updated dependencies [0bf34d9]
+- Updated dependencies [0dcf307]
+- Updated dependencies [9ac6aca]
+  - @paprika/helpers@2.3.0
+  - @paprika/tokens@2.0.0
+  - @paprika/raw-button@1.0.12
+  - @paprika/stylers@1.1.6
+
 ## 2.0.4-next.2
 
 ### Patch Changes