+++ packages/RawButton/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [0bf34d9]
+- Updated dependencies [9ac6aca]
+  - @paprika/tokens@2.0.0
+  - @paprika/stylers@1.1.6
+
 ## 1.0.12-next.1
 
 ### Patch Changes